consolidate gob registration
