+++ token/client.go
@@ -1,18 +1,10 @@
 package token
 
 import (
-	"encoding/gob"
 	"net/rpc"
 
 	"github.com/hashicorp/go-plugin"
 )
-
-func init() {
-	gob.Register(map[string]interface{}{})
-	gob.Register([]string{})
-	gob.Register(Token{})
-	gob.Register(plugin.BasicError{})
-}
 
 // Client implements the RPC client for token plugins.
 type Client struct {
