+++ token/pkg.go
@@ -1,0 +1,14 @@
+package token
+
+import (
+	"encoding/gob"
+
+	"github.com/hashicorp/go-plugin"
+)
+
+func init() {
+	gob.Register(map[string]interface{}{})
+	gob.Register([]string{})
+	gob.Register(Token{})
+	gob.Register(plugin.BasicError{})
+}
